% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-helperfunctions.R
\name{add_separators}
\alias{add_separators}
\title{Add seperators into an NHS number}
\usage{
add_separators(nhsnumber, separator = " ")
}
\arguments{
\item{nhsnumber}{nhsnumber as string}

\item{separator}{Character used to seperate digit-blocks. Default as space.}
}
\value{
nhs number as string with seperators after 3 and 6 characters
}
\description{
Adds seperator an NHS numer in the form 123 456 7890. Spaces are usually used to seperate digit-blocks
but other speators can be choosen
}
\examples{
add_separators("4000000004")
add_separators("4000000004",separator = "-")
}
