% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-helperfunctions.R
\name{calculate_check_digit}
\alias{calculate_check_digit}
\title{Calculate NHS Number check digit}
\source{
https://www.datadictionary.nhs.uk/data_dictionary/attributes/n/nhs/nhs_number_de.asp
}
\usage{
calculate_check_digit(nhsnumber)
}
\arguments{
\item{nhsnumber}{A 9 or 10 digit NHS Number. Parsed as a string.\cr If a 10 digit NHS number is inputed only the first 9 digits
are used to calcualte the check digit. If the number is valid the integer produced will match the 10th digit of the numbers}
}
\value{
The check digit as an integer. \cr \cr If the check digit is less than 10 the supplied \code{nhsnumber} is valid.
 A check digit of 10 means the supplied \code{nhsnumber} is invalid
}
\description{
Given the first 9 or 10 digits of a 10-digit NHS number, calculates what the check digit should be.
}
\details{
This procedure is calcualted by the method described in
 https://www.datadictionary.nhs.uk/data_dictionary/attributes/n/nhs/nhs_number_de.asp \cr
 \cr
 A summary of this method is:\cr
 1. Multiply each of the first nine digits by a weighting factor\cr
 2. Add the results of each multiplication together.\cr
 3. Divide the total by 11 and establish the remainder.\cr
 4. Subtract the remainder from 11 to give the check digit.\cr
 NB. Check digits of 11 are converted to 0.
}
\examples{
 calculate_check_digit("000000000") # Returns 0
 calculate_check_digit("411326996") #Returns 3


}
