% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-helperfunctions.R
\docType{data}
\name{checkdigit_weights}
\alias{checkdigit_weights}
\title{Data frame of weights given to each digit position when calcualting NHS number check digits}
\format{A data frame with 9 rows and 2 variables
\describe{
\item{digit.position}{indexed position from left to right in an NHS number}
\item{weight}{the weight assinged to that value in the check digit alogoritm}
}}
\source{
https://www.datadictionary.nhs.uk/data_dictionary/attributes/n/nhs/nhs_number_de.asp
}
\usage{
checkdigit_weights
}
\description{
NHS numbers are 9 digits followed by a check digit. The check digit is calculated by an algorithim
that associates each digit by its \code{digit.position} -from left to right- with a \code{weight}. The digit positions
and their associated weights are shown in this data frame.
}
\keyword{datasets}
