% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_nhs_number}
\alias{generate_nhs_number}
\title{Generate a sequence of NHS numbers}
\usage{
generate_nhs_number(n, determ = FALSE, seperators = NA,
  nhsnumber_value_range = NULL)
}
\arguments{
\item{n}{Integer. The number of NHS numbers to generate}

\item{determ}{Logical.\cr
If \code{TRUE} returns a predictable sequence of 10-digit NHS numbers starting at the lower bound of the specified range. See
 - \code{\link{deterministic_nhs_number}} for a detailed description.\cr
If \code{FALSE} generates NHS numbers at random within the the specified range-
see \code{\link{random_nhs_number}} for more details. \cr Default is \code{FALSE}.}

\item{seperators}{Character. Add the symbol (in " "'s) that is wanted as the seperator. Seperators are added
e.g. seperators = "-" : XXX-XXX-XXXX}

\item{nhsnumber_value_range}{Numeric range in which the sequence NHS numbers is derived.\cr
Input as a vector detailing the limits of the range (inclusive) e.g. \code{c(lower.bound, upper.bound)}. For multiple non-consecutive ranges
input a list of vectors. \cr
Ranges must not be more than 9 digits long. \cr
When \code{range} is \code{NULL} (default) the sequence is constructed from the ranges currently issued in England, Wales and the Isle of Man.
That is 400000000-499999999 and 600000000-708800001.}
}
\value{
Generates a vector of \code{n} 10-digit NHS numbers
}
\description{
This function generates 10-digit NHS numbers, which are the unique patient identifier used by the National
Health Service. \cr Note that, at some point point in the past, NHS numbers were 9-digits long- these cannot be gerenated
with this function
}
\examples{
generate_nhs_number(n=1000)
generate_nhs_number(n = 10, determ = FALSE)
generate_nhs_number(n = 10, determ = TRUE)
generate_nhs_number(n = 10, determ = FALSE, seperators = " ")
generate_nhs_number(n = 10, determ = TRUE, seperators = " ")
generate_nhs_number(n = 10, determ = TRUE, nhsnumber_value_range = c(100000000,200000000))
generate_nhs_number(n = 10, nhsnumber_value_range = c(100000000,200000000))
}
\seealso{
\code{\link{random_nhs_number}}\cr
\code{\link{deterministic_nhs_number}}
}
