% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_nhs_number_test}
\alias{generate_nhs_number_test}
\title{Randomly generate and NHS number test example version}
\usage{
generate_nhs_number_test(parameter)
}
\arguments{
\item{parameter}{A number.}
}
\value{
A random sequence of digits of length paramter
}
\description{
This randomly generates a random sequence of numbers of length \code{parameter}
}
\details{
This a test version a simply makes an \code{parameter} digit random number. It is used in testing so should be retained
}
\examples{
generate_nhs_number(8)
generate_nhs_number(2)

}
