% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{is_valid_nhs_number}
\alias{is_valid_nhs_number}
\title{Check an NHS number is valid}
\usage{
is_valid_nhs_number(nhsnumber, show_warnings = TRUE)
}
\arguments{
\item{nhsnumber}{potential nhs number to be validated. Input as string.}

\item{show_warnings}{If \code{TRUE} (default) warnings explaning why the NHS number is printed. If \code{FALSE}
these warnings are not displayed.}
}
\value{
Logical TRUE/FALSE. IF FALSE a warning will be displayed indicating the reason for the faliure.
}
\description{
This function validates potential NHS numbers- checking if they fit the requriments of an NHS number.\cr
\code{is_valid_nhsnumber} validates several features of the ptoetial NHS number (input as string) to
check that it is valid. THe function checks:\cr
1. That the input value is a character\cr
2. That the input string is 10 characters long.\cr
3. The input string contains only numbers (0-9).\cr
4. The check digit (calcualted from the first 9 digits) is not 10 (incidating an invalid NHS number) \cr
5. The 10th check digit is correct (matches that calcualted independtly via the published algorithin).\cr
\cr
If all these requirements are met the funtion return \code{TRUE}, if any fail it returns \code{FALSE}
}
\examples{
is_valid_nhs_number("0000000000") # TRUE
is_valid_nhs_number("0000000001") # FALSE
is_valid_nhs_number("0000r00000") # FALSE
is_valid_nhs_number("000000000") # FALSE


}
