% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsnumbergenerator-package.R
\docType{package}
\name{nhsnumbergenerator}
\alias{nhsnumbergenerator}
\alias{nhsnumbergenerator-package}
\title{nhsnumbergenerator: a package for generatring nhs numbers}
\description{
The pacakge contains functions to generate nhs numbers
}
\section{nhsnumbergenerator functions}{

More description
}

\seealso{
Useful links:
\itemize{
  \item \url{}
}

}
\keyword{internal}
