% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-additionalfunctions.R
\name{nine_digitify}
\alias{nine_digitify}
\title{Convert a string-number to have 9 digits}
\usage{
nine_digitify(nhsnumber)
}
\arguments{
\item{nhsnumber}{a potential NHS number or vector of potential NHS numbers.}
}
\value{
Nine-digit potential NHS number as string
}
\description{
NHS numbers must have nine digits. This function coverts numbers with fewer digits to having nine digits
by prefixing 0's. Numbers already with nine digits are unchanged. Numbers with more than nine digits give
an error.
}
\examples{
nine_digitify("01")
nine_digitify(c("01","02"))
nine_digitify("000000001") # returns unchanged
}
