% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-helperfunctions.R
\name{random_nhs_number}
\alias{random_nhs_number}
\title{Generate NHS numbers in an unpredictable sequence}
\usage{
random_nhs_number(n, nhsnums.range = NULL)
}
\arguments{
\item{n}{number of nhs numbers to return}

\item{range}{Numeric range in which the sequence NHS numbers is derived.\cr
Input as a vector detailing the limits of the range (inclusive) e.g. \code{c(lower.bound, upper.bound)}. For multiple non-consecutive ranges
input a list of vectors. \cr
Ranges must not be more than 9 digits long. \cr
When \code{range} is \code{NULL} (default) the sequence is constructed from the ranges currently issued in England, Wales and the Isle of Man.
That is 400000000-499999999 and 600000000-708800001.}
}
\value{
A vector of nhs number strings
}
\description{
Returns a generator for an unpredictable sequence of 10-digit NHS numbers.
}
\details{
\code{deterministic_nhs_number} produces
NHS numbers in an ascending sequence until n nhs numbers are reached. In contrast, \code{random_nhs_number}
generates nhsnumbers at random in the range specified until n is reached.

The default ranges are the ones currently issued in England, Wales and the Isle of Man.  Numbers outside of this
range may be valid but could conflict with identifiers used in Northern Ireland and Scotland.
See https://en.wikipedia.org/wiki/NHS_number
}
\examples{
random_nhs_number(10)
random_nhs_number(10, nhsnums.range = c(100000000,200000000))
random_nhs_number(10, nhsnums.range = list(c(100000000,200000000),c(300000000,400000000)))
}
