% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-helperfunctions.R
\name{remove_separators}
\alias{remove_separators}
\title{Remove separators from an NHS number}
\usage{
remove_separators(nhsnumber, suppress_warnings = TRUE)
}
\arguments{
\item{nhsnumber}{nhsnumber as string}

\item{suppress_warnings}{Logical. Warnings of the incorrect digit numbers are or are not supressed. Default is TRUE}
}
\value{
nhs number as string with separators after 3 and 6 characters
}
\description{
Removes all characters from a potential NHS number that are not numbers (0-9).
}
\details{
A warning is produced if the output has the wrong number of digits to be a valid
NHS number (10).
}
\examples{
remove_separators("400 000 0004")
remove_separators("400-000-0004")
}
