% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-additionalfunctions.R
\name{sample_from_ranges}
\alias{sample_from_ranges}
\title{Samples from two or more ranges}
\usage{
sample_from_ranges(listofranges)
}
\arguments{
\item{listofranges}{Ranges from which a value should be selected. Each list item should be a vector of the
form \code{c(lowerbound,upperbound)}.}
}
\value{
a numeric value selected at random from the input ranges
}
\description{
Selects a single value at random from two or more ranges.
}
